tests/wpt/tests/css/css-gaps/parsing/rule-outset-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Gap Decorations: rule-outset shorthands</title>
<link rel="help" href="https://www.w3.org/TR/css-gaps-1/#propdef-rule-outset">
<meta name="assert"
    content="rule-outset supports the full grammar '<length-percentage> <length-percentage>? /[<length-percentage> <length-percentage>?]?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
    <script>
        const rule_properties = {
            'column-rule-outset': ['column-rule-edge-start-outset', 'column-rule-edge-end-outset',
                'column-rule-interior-start-outset', 'column-rule-interior-end-outset'],
            'row-rule-outset': ['row-rule-edge-start-outset', 'row-rule-edge-end-outset',
                'row-rule-interior-start-outset', 'row-rule-interior-end-outset']
        };
        const testCases = [
            {
                input: '0px',
                expected: {
                    ruleEdgeStartOutset: '0px',
                    ruleEdgeEndOutset: '0px',
                    ruleInteriorStartOutset: '0px',
                    ruleInteriorEndOutset: '0px',
                }
            },
            {
                input: '10px',
                expected: {
                    ruleEdgeStartOutset: '10px',
                    ruleEdgeEndOutset: '10px',
                    ruleInteriorStartOutset: '10px',
                    ruleInteriorEndOutset: '10px'
                }
            },
            {
                input: '50%',
                expected: {
                    ruleEdgeStartOutset: '50%',
                    ruleEdgeEndOutset: '50%',
                    ruleInteriorStartOutset: '50%',
                    ruleInteriorEndOutset: '50%',
                }
            },
            {
                input: '10px 20px',
                expected: {
                    ruleEdgeStartOutset: '10px',
                    ruleEdgeEndOutset: '20px',
                    ruleInteriorStartOutset: '10px',
                    ruleInteriorEndOutset: '20px',
                }
            },
            {
                input: '10px 20px / 30px',
                expected: {
                    ruleEdgeStartOutset: '10px',
                    ruleEdgeEndOutset: '20px',
                    ruleInteriorStartOutset: '30px',
                    ruleInteriorEndOutset: '30px',
                }
            },
            {
                input: '10px / 20px 30px',
                expected: {
                    ruleEdgeStartOutset: '10px',
                    ruleEdgeEndOutset: '10px',
                    ruleInteriorStartOutset: '20px',
                    ruleInteriorEndOutset: '30px',
                }
            },
            {
                input: '10px 20px / 30px 40px',
                expected: {
                    ruleEdgeStartOutset: '10px',
                    ruleEdgeEndOutset: '20px',
                    ruleInteriorStartOutset: '30px',
                    ruleInteriorEndOutset: '40px',
                }
            },
        ];


        for (rule_property in rule_properties) {
            const test_cases = testCases;

            for (const { input, expected } of test_cases) {
                const [ruleEdgeStartOutset, ruleEdgeEndOutset, ruleInteriorStartOutset, ruleInteriorEndOutset] = rule_properties[rule_property];
                test_shorthand_value(rule_property, input, {
                    [ruleEdgeStartOutset]: expected.ruleEdgeStartOutset,
                    [ruleEdgeEndOutset]: expected.ruleEdgeEndOutset,
                    [ruleInteriorStartOutset]: expected.ruleInteriorStartOutset,
                    [ruleInteriorEndOutset]: expected.ruleInteriorEndOutset,
                });
            }
        }
    </script>
</body>
</html>
